refactor(embedded): extract entrypoint resolution from dispatcher.
